refactor: move all the request reading functions to request.go
